not-working
This code creates a simple web form where users can enter the title, description and location of the food related illness. When the form is submitted, it sends a POST request to the Flask server, which creates a new report in the database.

Please note that this is a simplified version of the application, meant to demonstrate the basics of how it could be built.